+++ apps/editor-dev/pages/base-editor.jsx
@@ -1,18 +1,26 @@
-import {
-  BaseEditor,
-  InjectView,
-  useCommand,
-  useSetConfig,
-} from "@elf-framework/base-editor";
+import { BaseEditor } from "@elf-framework/base-editor";
 import { start } from "@elf-framework/sapa";
-import { AppLayout, AppLayoutItem } from "@elf-framework/ui";
 import "@elf-framework/ui/style.css";
+
+import { CenterPlugin } from "./plugins/CenterPlugin";
+import { LeftPlugin } from "./plugins/LeftPlugin";
+import { RenderPlugin } from "./plugins/RenderPlugin";
+import { RightPlugin } from "./plugins/RightPlugin";
+import { ToolbarPlugin } from "./plugins/ToolbarPlugin";
 
 start(function () {
   return (
     <div>
       <BaseEditor
+        configs={{
+          key: "value",
+        }}
         plugins={[
+          ToolbarPlugin,
+          RightPlugin,
+          LeftPlugin,
+          CenterPlugin,
+
           async function (editor) {
             editor.registerConfig({
               key: "yellow",
@@ -29,69 +37,15 @@ start(function () {
                 return 10;
               },
             });
+
+            console.log(editor.configs.get("yellow"));
+
+            
           },
 
-          function (editor) {
-            editor.registerUI({
-              view: () => {
-                return (
-                  <button
-                    type="button"
-                    onClick={async () => {
-                      const ret = await useCommand("my-command");
-
-                      console.log("return", ret);
-
-                      useSetConfig("yellow", "red" + Math.random());
-                    }}
-                  >
-                    Sample
-                  </button>
-                );
-              },
-            });
-
-            editor.registerGroupUI("right-panel", {
-              view: () => {
-                return (
-                  <div>
-                    <button
-                      type="button"
-                      onClick={async () => {
-                        const ret = await useCommand("my-command");
-
-                        console.log("return", ret);
-                      }}
-                    >
-                      Sample
-                    </button>
-                  </div>
-                );
-              },
-            });
-          },
+          RenderPlugin,
         ]}
-      >
-        <AppLayout>
-          <AppLayoutItem direction="top" height={40}>
-            <InjectView views={["top-panel"]} />
-          </AppLayoutItem>
-          <AppLayoutItem
-            direction="left"
-            width={280}
-            maxWidth={350}
-            resizable={true}
-          >
-            <InjectView views={["left-panel"]} />
-          </AppLayoutItem>
-          <AppLayoutItem direction="right" width={300}>
-            <InjectView groups={["right-panel"]} />
-          </AppLayoutItem>
-          <AppLayoutItem direction="center">
-            <InjectView views={["center-panel"]} />
-          </AppLayoutItem>
-        </AppLayout>
-      </BaseEditor>
+      ></BaseEditor>
     </div>
   );
 });