+++ packages/sapa/src/App.js
@@ -2,7 +2,7 @@ import { Dom } from "./functions/Dom";
 import { registRootElementInstance } from "./functions/registElement";
 import { VNode } from "./functions/vnode";
 // import { Router } from "./Router";
-import { UIElement } from "./UIElement";
+import { createComponentInstance } from "./UIElement";
 
 /**
  * UIElement 렌더링 하기
@@ -13,18 +13,14 @@ import { UIElement } from "./UIElement";
 export const start = (ElementClass, opt = {}) => {
   const $container = Dom.create(opt.container || document.body);
 
-  console.log($container.children());
-
   const $targetElement = $container.children().find((it) => it.el.__component);
-
-  console.log($targetElement);
 
   if (ElementClass instanceof VNode) {
     const rootVNode = ElementClass;
     ElementClass = () => rootVNode;
   }
 
-  const app = UIElement.createElementInstance(ElementClass, opt);
+  const app = createComponentInstance(ElementClass, null, opt);
 
   // hmr 로 인해서 동일한 패턴으로 start 가 되었을 때
   // dom 이 살아있지 않으면 추가
@@ -38,10 +34,12 @@ export const start = (ElementClass, opt = {}) => {
     app.render($container);
   }
   app.$el.el.__component = app;
-  registRootElementInstance(app);
+  registRootElementInstance(app, $container);
 
   return app;
 };
+
+export const render = start;
 
 export async function renderToHtml(ElementClass, opt) {
   if (ElementClass instanceof VNode) {
@@ -49,7 +47,7 @@ export async function renderToHtml(ElementClass, opt) {
     ElementClass = () => rootVNode;
   }
 
-  const app = UIElement.createElementInstance(ElementClass, opt);
+  const app = createComponentInstance(ElementClass, null, opt);
   const html = await app.renderToHtml();
 
   return html;